pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Me"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.15.1"
psycopg2 = "^2.9.9"
fastapi = "^0.115.10"
uvicorn = "^0.34.0"
gunicorn = "^23.0.0"
sqlalchemy = "2.0.38"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.8.1"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = "^1.7.4"
bcrypt = "^4.3.0"
argon2-cffi = "^23.1.0"
python-multipart = "^0.0.20"
pyyaml = "^6.0.1"
email-validator = "^2.1.1"
sqlmodel = "^0.0.23"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
requests = "^2.31.0"
httpx = "^0.27.0"
sqlalchemy-utils = "^0.41.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
